# application/static/swagger.yaml
swagger: "2.0"
info:
  title: "My Mechanic Shop API"
  description: "Flask API for customers, mechanics, service tickets, and inventory."
  version: "1.0.0"
host: "localhost:5000"
schemes:
  - "http"
consumes:
  - "application/json"
produces:
  - "application/json"

# ---- AUTH (Bearer Token) ----
securityDefinitions:
  bearerAuth:
    type: apiKey
    name: Authorization
    in: header
    description: "Use: Bearer <auth_token>"
paths:
...
  /customers/:
    post:
      tags: [Customers]
      summary: "Create customer"
      description: "Register a new customer. Email must be unique."
      parameters:
        - in: body
          name: body
          required: true
          schema: { $ref: "#/definitions/CustomerCreatePayload" }
      responses:
        201:
          description: "Created"
          schema: { $ref: "#/definitions/CustomerResponse" }
          examples:
            application/json:
              id: 1
              name: "John Doe"
              email: "john.doe@example.com"
              phone: "1234567890"
        400:
          description: "Validation error or email already exists"
          schema: { $ref: "#/definitions/ErrorMessage" }
          examples:
            application/json:
              error: "Bad Request"
              message: "Email already exists"

    get:
      tags: [Customers]
      summary: "List customers (paginated)"
      description: "Supports query params: limit (1-100), offset (>=0)."
      parameters:
        - name: limit
          in: query
          type: integer
          required: false
        - name: offset
          in: query
          type: integer
          required: false
      responses:
        200:
          description: "OK"
          schema: { $ref: "#/definitions/CustomersListResponse" }
          examples:
            application/json:
              customers:
                - id: 1
                  name: "Jane Doe"
                  email: "jane.doe@example.com"
                  phone: "123-456-7890"
                - id: 2
                  name: "John Smith"
                  email: "john.smith@example.com"
                  phone: "098-765-4321"
              limit: 10
              offset: 0
              total: 2
        400:
          description: "Bad pagination values"
          schema: { $ref: "#/definitions/ErrorMessage" }
          examples:
            application/json:
              error: "Bad Request"
              message: "Invalid limit or offset"
  /customers/login:
    post:
      tags: [Customers]
      summary: "Login (returns auth token)"
      description: "Returns a JWT token for authentication."
      parameters:
        - in: body
          name: body
          required: true
          schema: { $ref: "#/definitions/CustomerLoginPayload" }
      responses:
        200:
          description: "OK"
          schema: { $ref: "#/definitions/CustomerLoginResponse" }
          examples:
            application/json:
              auth_token: "1234567890..."
              customer_id: 1
        401:
          description: "Invalid credentials"
          schema: { $ref: "#/definitions/ErrorMessage" }
          examples:
            application/json:
              error: "Unauthorized"
              message: "Invalid email or password."
  /customers/my-tickets:
    get:
      tags: [Customers]
      summary: "My tickets (auth)"
      description: "Returns a list of tickets for the logged-in customer."
      security:
        - bearerAuth: []
      responses:
        200:
          description: "OK"
          schema:
            type: array
            items: { $ref: "#/definitions/TicketResponse" }
          examples:
            application/json:
              - id: 10
                VIN: "1HGCM82633A004352"
                service_date: "2025-02-18"
                service_desc: "Oil change"
                customer_id: 1
        401:
          description: "Missing/invalid token"
          schema: { $ref: "#/definitions/ErrorMessage" }
          examples:
            application/json:
              error: "Unauthorized"
              message: "Missing/invalid token"
  /customers/me:
    put:
      tags: [Customers]
      summary: "Update logged-in customer (auth)"
      description: "Updates the logged-in customer's information."
      security:
        - bearerAuth: []
      parameters:
        - in: body
          name: body
          required: true
          schema: { $ref: "#/definitions/CustomerCreatePayload" }
      responses:
        200:
          description: "OK"
          schema: { $ref: "#/definitions/CustomerResponse" }
          examples:
            application/json:
              id: 1
              name: "John Doe Updated"
              email: "john.doe.updated@example.com"
              phone: "2345678910"
        401:
          description: "Missing/invalid token"
          schema: { $ref: "#/definitions/ErrorMessage" }
          examples:
            application/json:
              error: "Unauthorized"
              message: "Missing/invalid token"

    delete:
      tags: [Customers]
      summary: "Delete logged-in customer (auth)"
      description: "Deletes the logged-in customer."
      security:
        - bearerAuth: []
      responses:
        204:
          description: "No Content (customer deleted successfully)"
        401:
          description: "Missing/invalid token"
          schema: { $ref: "#/definitions/ErrorMessage" }
          examples:
            application/json:
              error: "Unauthorized"
              message: "Missing/invalid token"
  /customers/{customer_id}:
    get:
      tags: [Customers]
      summary: "Get customer by id"
      description: "Returns a customer by their ID."
      parameters:
        - name: customer_id
          in: path
          required: true
          type: integer
      responses:
        200:
          description: "OK"
          schema: { $ref: "#/definitions/CustomerResponse" }
          examples:
            application/json:
              id: 1
              name: "John Doe"
              email: "john.doe@example.com"
              phone: "1234567890"
        404:
          description: "Not found"
          schema: { $ref: "#/definitions/ErrorMessage" }
          examples:
            application/json:
              error: "Not Found"
              message: "Customer not found"

  # -------------------- Mechanics --------------------
...
definitions:
...
  ErrorMessage:
    type: object
    properties:
      error:
        type: string
      message:
        type: string

  # ---- Customers ----
  CustomerCreatePayload:
    type: object
    required: [name, email, password]
    properties:
      name: { type: string }
      email: { type: string }
      phone: { type: string }
      password: { type: string }
  CustomerResponse:
    type: object
    properties:
      id: { type: integer }
      name: { type: string }
      email: { type: string }
      phone: { type: string }
  CustomerLoginPayload:
    type: object
    required: [email, password]
    properties:
      email: { type: string }
      password: { type: string }
  CustomerLoginResponse:
    type: object
    properties:
      status: { type: string }
      message: { type: string }
      auth_token: { type: string }
  CustomersListResponse:
    type: object
    properties:
      limit: { type: integer }
      offset: { type: integer }
      count: { type: integer }
      customers:
        type: array
        items:
          $ref: "#/definitions/CustomerResponse"
...
  TicketResponse:
...